requirements.txt
asgiref==3.5.2
# Django==4.1.3
Django==3.2.16
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
requests==2.28.1
tweepy==4.12.1
psycopg2==2.9.5
django-allauth==0.51.0
nltk==3.7
matplotlib==3.6.2
scikit-learn==1.1.3
django-crontab==0.7.1
djangorestframework==3.14.0
django-plotly-dash==2.1.3
channels==4.0.0
daphne==4.0.0
redis==4.4.0
django-redis==5.2.0
channels-redis==4.0.0
dash-renderer==1.9.1
dash-bootstrap-components==1.2.1

#jupyterlab==3.5.1